Moved 2 functions from player to game_data
These functions were cause a circular import error, so I moved them to game_data.py with the same functionality.
Also, game board was generating the same Space() object over and over for the board state, so I changed that.
Also, x and y were switched for a few loops so I changed that